New edit clipboard system & new variantToMimeData
* Changed the whole function `Clip::variantToMimeData()` to match the correct enum values
* Changed the keys of `history` to milliseconds since epoch (`qint64`)
* Removed the `editClipboard()` method in `MainWindow` because it is not generic and not adapted to POO. Instead, I created an absatract class `EditClipboard` and its children to edit the clipboard using different functions. However, I cannot compile with the current version (Maybe I should try to clean the whole project and then compile again)

+++ clip.h
@@ -18,6 +18,7 @@
 #include <QMetaType>
 #include <QFont>
 #include <QRegularExpression>
+#include <QUuid>
 
 using namespace std;
 
@@ -108,8 +109,8 @@ public:
 	
 	/* GETTER & SETTERS */
 	
-	QMap<uint, QVariant>* getHistory();
-	void setHistory(const QMap<uint, QVariant>& value);
+	QMap<qint64, QVariant>* getHistory();
+	void setHistory(const QMap<qint64, QVariant>& value);
 	
 signals:
 	void modeChanged(QClipboard::Mode);
@@ -131,7 +132,7 @@ private slots:
 private:
 	static Clip* clip;
 	QClipboard* clipboard = QGuiApplication::clipboard();
-	QMap<uint, QVariant> history;
+	QMap<qint64, QVariant> history;
 	
 	static QRegularExpression urlPattern;
 };
